Refactored swappable debug profile methods into composable pieces
