+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/api/profiles/swappable/events/SwappableEventProfile.scala
@@ -1,0 +1,22 @@
+package org.senkbeil.debugger.api.profiles.swappable.events
+
+import org.senkbeil.debugger.api.lowlevel.JDIArgument
+import org.senkbeil.debugger.api.lowlevel.events.EventType.EventType
+import org.senkbeil.debugger.api.pipelines.Pipeline.IdentityPipeline
+import org.senkbeil.debugger.api.profiles.swappable.SwappableDebugProfile
+import org.senkbeil.debugger.api.profiles.traits.events.EventProfile
+
+/**
+ * Represents a swappable profile for events that redirects the
+ * invocation to another profile.
+ */
+trait SwappableEventProfile extends EventProfile {
+  this: SwappableDebugProfile =>
+
+  override def onEventWithData(
+    eventType: EventType,
+    extraArguments: JDIArgument*
+  ): IdentityPipeline[EventAndData] = {
+    withCurrentProfile.onEventWithData(eventType, extraArguments: _*)
+  }
+}
