+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/api/profiles/swappable/monitors/SwappableMonitorContendedEnteredProfile.scala
@@ -1,0 +1,20 @@
+package org.senkbeil.debugger.api.profiles.swappable.monitors
+
+import org.senkbeil.debugger.api.lowlevel.JDIArgument
+import org.senkbeil.debugger.api.pipelines.Pipeline.IdentityPipeline
+import org.senkbeil.debugger.api.profiles.swappable.SwappableDebugProfile
+import org.senkbeil.debugger.api.profiles.traits.monitors.MonitorContendedEnteredProfile
+
+/**
+ * Represents a swappable profile for monitor contended entered events that
+ * redirects the invocation to another profile.
+ */
+trait SwappableMonitorContendedEnteredProfile extends MonitorContendedEnteredProfile {
+  this: SwappableDebugProfile =>
+
+  override def onMonitorContendedEnteredWithData(
+    extraArguments: JDIArgument*
+  ): IdentityPipeline[MonitorContendedEnteredEventAndData] = {
+    withCurrentProfile.onMonitorContendedEnteredWithData(extraArguments: _*)
+  }
+}
