+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/api/profiles/swappable/monitors/SwappableMonitorWaitProfile.scala
@@ -1,0 +1,20 @@
+package org.senkbeil.debugger.api.profiles.swappable.monitors
+
+import org.senkbeil.debugger.api.lowlevel.JDIArgument
+import org.senkbeil.debugger.api.pipelines.Pipeline.IdentityPipeline
+import org.senkbeil.debugger.api.profiles.swappable.SwappableDebugProfile
+import org.senkbeil.debugger.api.profiles.traits.monitors.MonitorWaitProfile
+
+/**
+ * Represents a swappable profile for monitor wait events that
+ * redirects the invocation to another profile.
+ */
+trait SwappableMonitorWaitProfile extends MonitorWaitProfile {
+  this: SwappableDebugProfile =>
+
+  override def onMonitorWaitWithData(
+    extraArguments: JDIArgument*
+  ): IdentityPipeline[MonitorWaitEventAndData] = {
+    withCurrentProfile.onMonitorWaitWithData(extraArguments: _*)
+  }
+}
