+++ scala-debugger-api/src/main/scala/org/senkbeil/debugger/api/profiles/swappable/watchpoints/SwappableAccessWatchpointProfile.scala
@@ -1,0 +1,36 @@
+package org.senkbeil.debugger.api.profiles.swappable.watchpoints
+
+import org.senkbeil.debugger.api.lowlevel.JDIArgument
+import org.senkbeil.debugger.api.pipelines.Pipeline.IdentityPipeline
+import org.senkbeil.debugger.api.profiles.swappable.SwappableDebugProfile
+import org.senkbeil.debugger.api.profiles.traits.watchpoints.AccessWatchpointProfile
+
+/**
+ * Represents a swappable profile for access watchpoint events that redirects
+ * the invocation to another profile.
+ */
+trait SwappableAccessWatchpointProfile extends AccessWatchpointProfile {
+  this: SwappableDebugProfile =>
+
+  override def onAccessFieldWatchpointWithData(
+    className: String,
+    fieldName: String,
+    extraArguments: JDIArgument*
+  ): IdentityPipeline[AccessWatchpointEventAndData] = {
+    withCurrentProfile.onAccessFieldWatchpointWithData(
+      className,
+      fieldName,
+      extraArguments: _*
+    )
+  }
+
+  override def onAccessInstanceWatchpointWithData(
+    instanceVarName: String,
+    extraArguments: JDIArgument*
+  ): IdentityPipeline[AccessWatchpointEventAndData] = {
+    withCurrentProfile.onAccessInstanceWatchpointWithData(
+      instanceVarName,
+      extraArguments: _*
+    )
+  }
+}
