+++ scala-debugger-api/src/test/scala/org/senkbeil/debugger/api/profiles/swappable/classes/SwappableClassUnloadProfileSpec.scala
@@ -1,0 +1,44 @@
+package org.senkbeil.debugger.api.profiles.swappable.classes
+
+import org.scalamock.scalatest.MockFactory
+import org.scalatest.{FunSpec, Matchers, OneInstancePerTest}
+import org.senkbeil.debugger.api.lowlevel.JDIArgument
+import org.senkbeil.debugger.api.profiles.ProfileManager
+import org.senkbeil.debugger.api.profiles.swappable.SwappableDebugProfile
+import org.senkbeil.debugger.api.profiles.traits.DebugProfile
+
+class SwappableClassUnloadProfileSpec extends FunSpec with Matchers
+  with OneInstancePerTest with MockFactory
+{
+  private val mockDebugProfile = mock[DebugProfile]
+  private val mockProfileManager = mock[ProfileManager]
+
+  private val swappableDebugProfile = new Object with SwappableDebugProfile {
+    override protected val profileManager: ProfileManager = mockProfileManager
+  }
+
+  describe("SwappableClassUnloadProfile") {
+    describe("#onClassUnloadWithData") {
+      it("should invoke the method on the underlying profile") {
+        val arguments = Seq(mock[JDIArgument])
+
+        (mockProfileManager.retrieve _).expects(*)
+          .returning(Some(mockDebugProfile)).once()
+
+        (mockDebugProfile.onClassUnloadWithData _).expects(arguments).once()
+
+        swappableDebugProfile.onClassUnloadWithData(arguments: _*)
+      }
+
+      it("should throw an exception if there is no underlying profile") {
+        val arguments = Seq(mock[JDIArgument])
+
+        (mockProfileManager.retrieve _).expects(*).returning(None).once()
+
+        intercept[AssertionError] {
+          swappableDebugProfile.onClassUnloadWithData(arguments: _*)
+        }
+      }
+    }
+  }
+}
