+++ scala-debugger-api/src/test/scala/org/senkbeil/debugger/api/profiles/swappable/vm/SwappableVMDeathProfileSpec.scala
@@ -1,0 +1,44 @@
+package org.senkbeil.debugger.api.profiles.swappable.vm
+
+import org.scalamock.scalatest.MockFactory
+import org.scalatest.{FunSpec, Matchers, OneInstancePerTest}
+import org.senkbeil.debugger.api.lowlevel.JDIArgument
+import org.senkbeil.debugger.api.profiles.ProfileManager
+import org.senkbeil.debugger.api.profiles.swappable.SwappableDebugProfile
+import org.senkbeil.debugger.api.profiles.traits.DebugProfile
+
+class SwappableVMDeathProfileSpec extends FunSpec with Matchers
+  with OneInstancePerTest with MockFactory
+{
+  private val mockDebugProfile = mock[DebugProfile]
+  private val mockProfileManager = mock[ProfileManager]
+
+  private val swappableDebugProfile = new Object with SwappableDebugProfile {
+    override protected val profileManager: ProfileManager = mockProfileManager
+  }
+
+  describe("SwappableVMDeathProfile") {
+    describe("#onVMDeathWithData") {
+      it("should invoke the method on the underlying profile") {
+        val arguments = Seq(mock[JDIArgument])
+
+        (mockProfileManager.retrieve _).expects(*)
+          .returning(Some(mockDebugProfile)).once()
+
+        (mockDebugProfile.onVMDeathWithData _).expects(arguments).once()
+
+        swappableDebugProfile.onVMDeathWithData(arguments: _*)
+      }
+
+      it("should throw an exception if there is no underlying profile") {
+        val arguments = Seq(mock[JDIArgument])
+
+        (mockProfileManager.retrieve _).expects(*).returning(None).once()
+
+        intercept[AssertionError] {
+          swappableDebugProfile.onVMDeathWithData(arguments: _*)
+        }
+      }
+    }
+  }
+}
